Fix break in size

+++ benches/ops/read.rs
@@ -52,13 +52,17 @@ fn bench_read_full(c: &mut Criterion, op: Operator) {
         let path = uuid::Uuid::new_v4().to_string();
         let temp_data = TempData::generate(op.clone(), &path, content.clone());
 
-        group.throughput(criterion::Throughput::Bytes(size.bytes()));
+        group.throughput(criterion::Throughput::Bytes(size.bytes() as u64));
         group.bench_with_input(
             size.to_string(Base::Base2, Style::Abbreviated),
             &(op.clone(), &path),
             |b, (op, path)| {
                 b.to_async(&*TOKIO).iter(|| async {
-                    let r = op.object(path).range_reader(..=size.bytes()).await.unwrap();
+                    let r = op
+                        .object(path)
+                        .range_reader(..=size.bytes() as u64)
+                        .await
+                        .unwrap();
                     io::copy(r, &mut io::sink()).await.unwrap();
                 })
             },
@@ -84,10 +88,10 @@ fn bench_read_part(c: &mut Criterion, op: Operator) {
     ] {
         let content = gen_bytes(&mut rng, (size.bytes() * 2) as usize);
         let path = uuid::Uuid::new_v4().to_string();
-        let offset = size.bytes() / 2;
+        let offset = (size.bytes() / 2) as u64;
         let temp_data = TempData::generate(op.clone(), &path, content.clone());
 
-        group.throughput(criterion::Throughput::Bytes(size.bytes()));
+        group.throughput(criterion::Throughput::Bytes(size.bytes() as u64));
         group.bench_with_input(
             size.to_string(Base::Base2, Style::Abbreviated),
             &(op.clone(), &path),
@@ -118,12 +122,12 @@ fn bench_read_parallel(c: &mut Criterion, op: Operator) {
     ] {
         let content = gen_bytes(&mut rng, (size.bytes() * 2) as usize);
         let path = uuid::Uuid::new_v4().to_string();
-        let offset = size.bytes() / 2;
+        let offset = (size.bytes() / 2) as u64;
         let buf = vec![0; size.bytes() as usize];
         let temp_data = TempData::generate(op.clone(), &path, content.clone());
 
         for parallel in [1, 2, 4, 8, 16] {
-            group.throughput(criterion::Throughput::Bytes(parallel * size.bytes()));
+            group.throughput(criterion::Throughput::Bytes(parallel * size.bytes() as u64));
             group.bench_with_input(
                 format!(
                     "{}x{}",
@@ -138,7 +142,7 @@ fn bench_read_parallel(c: &mut Criterion, op: Operator) {
                                 let mut buf = buf.clone();
                                 let mut r = op
                                     .object(path)
-                                    .range_reader(offset..=offset + size.bytes())
+                                    .range_reader(offset..=offset + size.bytes() as u64)
                                     .await
                                     .unwrap();
                                 r.read_exact(&mut buf).await.unwrap();
